Extract `GraplStack` to a shared class
Essentially the same code was being used for both the Python and Rust
integration tests. Rather than duplicating things, we can now just use
a canonical implementation, ensuring access to anything from the
underlying `StackReference` that is needed.
